Add upper version bound for setInterruptSafety()
This function became a no-op [1] in RPM 4.18 where it was also
deprecated [2], and is completely gone in RPM 4.19 [3].  Adjust the
version check accordingly.

Note that the function itself seems to have no effect in RPM 4.16
already where the Berkeley DB backend was swapped for sqlite, but don't
risk breaking stuff here and just keep the call in place for those older
RPM versions.

More info here:

[1] https://github.com/rpm-software-management/rpm/commit/
    baeb71141744bb0db6bd72c580bfaf929b7b4fb2
[2] https://github.com/rpm-software-management/rpm/pull/1994
[3] https://github.com/rpm-software-management/rpm/pull/1992